fix(og): move param normalization out of route
